+++ lib/plex_utils/plex_utils.dart
@@ -32,9 +32,14 @@ double get _widthRatio {
   return WidgetsBinding.instance.platformDispatcher.views.first.physicalSize.width / WidgetsBinding.instance.platformDispatcher.views.first.devicePixelRatio;
 }
 
+isExtLargeScreen(BuildContext context) {
+  // return _widthRatio >= 1300;
+  return MediaQuery.of(context).size.width >= 1300;
+}
+
 isLargeScreen(BuildContext context) {
-  // return _widthRatio >= 900;
-  return MediaQuery.of(context).size.width >= 900;
+  // return _widthRatio >= 900 && _widthRatio < 1300;
+  return MediaQuery.of(context).size.width >= 900 && MediaQuery.of(context).size.width < 1300;
 }
 
 isMediumScreen(BuildContext context) {